requirements.txt
pyfiglet>=1.0.3
colorama>=0.4.6
pyautogui>=0.9.54
requests>=2.32.3
urllib3>=2.3.0
pyexecjs>=1.5.1
fake-useragent>=2.2.0